eatgo-customer-api/src/main/java/walter/study/restaurant/controller/UserController.java
package walter.study.restaurant.controller;

import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;
import walter.study.restaurant.application.UserService;
import walter.study.restaurant.domain.User;

import java.net.URI;
import java.net.URISyntaxException;

@RequiredArgsConstructor
@RestController
public class UserController {

    private final UserService userService;

    @PostMapping("/users")
    public ResponseEntity<?> create(@RequestBody User resource) throws URISyntaxException {

        String email = resource.getEmail();
        String name = resource.getName();
        String password = resource.getPassword();

        String url = "/users/" + resource.getId();

        User newUser = userService.registerUser(email, name, password);

        return ResponseEntity.created(new URI(url)).body("{}");
    }
}
